MF-507 tidy up specflow steps.
