+++ src/SFA.DAS.Reservations.Api.AcceptanceTests/Steps/ChangeOfPartySteps.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net;
-using System.Threading.Tasks;
-using FluentAssertions;
-using Microsoft.AspNetCore.Mvc;
+﻿using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
-using NUnit.Framework;
 using SFA.DAS.Reservations.Api.Controllers;
 using SFA.DAS.Reservations.Api.Models;
 using TechTalk.SpecFlow;
@@ -49,35 +43,6 @@ namespace SFA.DAS.Reservations.Api.AcceptanceTests.Steps
             var controller = Services.GetService<ReservationsController>();
 
             TestResults.Result = await controller.Change(TestData.ReservationId, request);
-        }
-
-        [Then("an http status code of (.*) is returned")]
-        public void ThenAnHttpStatusCodeIsReturned(int httpStatusCode)
-        {
-            switch (httpStatusCode)
-            {
-                case (int)HttpStatusCode.BadRequest:
-                    var badRequestTypes = new List<Type>
-                    {
-                        typeof(BadRequestResult), 
-                        typeof(BadRequestObjectResult)
-                    };
-                    badRequestTypes.Contains(TestResults.Result.GetType()).Should().BeTrue(
-                        $"actual return type is {TestResults.Result.GetType()}");
-                    break;
-                case (int)HttpStatusCode.OK:
-                    var okTypes = new List<Type>
-                    {
-                        typeof(OkResult), 
-                        typeof(OkObjectResult)
-                    };
-                    okTypes.Contains(TestResults.Result.GetType()).Should().BeTrue(
-                        $"actual return is {TestResults.Result}");
-                    break;
-                default:
-                    Assert.Fail("http status code not supported");
-                    break;
-            }
         }
     }
 }
